feature: connections and related models now live in the platform namespace
