+++ app/ui/src/app/app.component.ts
@@ -14,10 +14,9 @@ import { Notification, NotificationEvent } from 'patternfly-ng';
 import { Observable } from 'rxjs/Observable';
 import { ModalService } from './common/modal/modal.service';
 import { NavigationService } from './common/navigation.service';
-import { UserService } from '@syndesis/ui/platform';
+import { UserService, User } from '@syndesis/ui/platform';
 import { ConfigService } from './config.service';
 import { log } from './logging';
-import { User } from './model';
 import { TestSupportService } from './store/test-support.service';
 import { TourService } from 'ngx-tour-ngx-bootstrap';
 import { NotificationService } from 'app/common/ui-patternfly/notification-service';
