+++ app/ui/src/app/connections/create-page/configure-fields/configure-fields.component.ts
@@ -1,15 +1,11 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { RouterStateSnapshot } from '@angular/router';
 import { FormGroup } from '@angular/forms';
-import {
-  DynamicFormControlModel,
-  DynamicFormService
-} from '@ng-dynamic-forms/core';
+import { DynamicFormControlModel, DynamicFormService } from '@ng-dynamic-forms/core';
 import { Subscription } from 'rxjs/Subscription';
 
+import { Connection, CanComponentDeactivate } from '@syndesis/ui/platform';
 import { CurrentConnectionService } from '../current-connection';
-import { Connection } from '../../../model';
-import { CanComponentDeactivate } from '@syndesis/ui/platform';
 import { ModalService } from '../../../common/modal/modal.service';
 import { ConnectionConfigurationService } from '../../common/configuration/configuration.service';
