+++ app/ui/src/app/connections/list/list.component.ts
@@ -9,10 +9,10 @@ import {
 
 import { NotificationType } from 'patternfly-ng';
 
+import { Connections, Connection } from '@syndesis/ui/platform';
 import { ModalService } from '../../common/modal/modal.service';
 import { ConnectionStore } from '../../store/connection/connection.store';
 import { log, getCategory } from '../../logging';
-import { Connections, Connection } from '../../model';
 import { NotificationService } from 'app/common/ui-patternfly/notification-service';
 
 const category = getCategory('Connections');
