+++ app/ui/src/app/index.ts
@@ -1,2 +1,1 @@
-export * from './app.component';
 export * from './app.module';
