+++ app/ui/src/app/integration/edit-page/current-flow.service.spec.ts
@@ -4,15 +4,12 @@ import { RequestOptions, BaseRequestOptions, Http } from '@angular/http';
 import { MockBackend } from '@angular/http/testing';
 import { RestangularModule } from 'ngx-restangular';
 
-import { CurrentFlow, FlowEvent } from './current-flow.service';
-import { EventsService, IntegrationStore, IntegrationService } from '@syndesis/ui/store';
-import {
-  Connection,
-  Action,
-  TypeFactory
-} from '@syndesis/ui/model';
-import { createIntegration, createStep, Integration, Step, Steps } from '@syndesis/ui/integration';
+import {  Connection, Action } from '@syndesis/ui/platform';
 import { SyndesisCommonModule } from '@syndesis/ui/common';
+import { TypeFactory } from '@syndesis/ui/model';
+import { EventsService, IntegrationStore, IntegrationService } from '@syndesis/ui/store';
+import { createIntegration, createStep, Integration, Step, Steps } from '@syndesis/ui/integration';
+import { CurrentFlow, FlowEvent } from './current-flow.service';
 
 describe('CurrentFlow', () => {
   beforeEach(() => {
