+++ app/ui/src/app/integration/edit-page/flow-view/flow-view-step.component.ts
@@ -4,12 +4,12 @@ import { Observable } from 'rxjs/Observable';
 import { Subscription } from 'rxjs/Subscription';
 import { PopoverDirective } from 'ngx-bootstrap/popover';
 
-import { StepStore } from '@syndesis/ui/store';
-import { ChildAwarePage, CurrentFlow, FlowEvent } from '@syndesis/ui/integration/edit-page';
-import { log, getCategory } from '@syndesis/ui/logging';
-import { Integration, Step } from '@syndesis/ui/integration';
-import { Action } from '@syndesis/ui/model';
+import { Action } from '@syndesis/ui/platform';
 import { ModalService } from '@syndesis/ui/common';
+import { log, getCategory } from '@syndesis/ui/logging';
+import { StepStore } from '@syndesis/ui/store';
+import { Integration, Step } from '@syndesis/ui/integration';
+import { ChildAwarePage, CurrentFlow, FlowEvent } from '@syndesis/ui/integration/edit-page';
 
 const category = getCategory('IntegrationsCreatePage');
