+++ app/ui/src/app/integration/edit-page/list-actions/list-actions.component.ts
@@ -1,7 +1,7 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 
-import { Actions, Action } from '@syndesis/ui/model';
+import { Actions, Action } from '@syndesis/ui/platform';
 import { log, getCategory } from '@syndesis/ui/logging';
 
 const category = getCategory('Actions');
