+++ app/ui/src/app/integration/edit-page/step-configure/filter-steps/basic-filter.component.ts
@@ -1,25 +1,13 @@
-import {
-  Component,
-  EventEmitter,
-  Input,
-  Output,
-  ViewEncapsulation,
-  OnChanges,
-} from '@angular/core';
+import { Component, EventEmitter, Input, Output, ViewEncapsulation, OnChanges } from '@angular/core';
 import { FormGroup, FormControl, FormArray } from '@angular/forms';
-import {
-  DynamicFormService,
-  DynamicFormControlModel,
-  DynamicFormArrayModel,
-  DynamicInputModel
-} from '@ng-dynamic-forms/core';
+import { DynamicFormService, DynamicFormControlModel, DynamicFormArrayModel, DynamicInputModel } from '@ng-dynamic-forms/core';
 
+import { DataShape } from '@syndesis/ui/platform';
+import { log, getCategory } from '@syndesis/ui/logging';
+import { DATA_MAPPER } from '@syndesis/ui/store';
 import { CurrentFlow, FlowEvent } from '@syndesis/ui/integration/edit-page';
 import { IntegrationSupportService } from '../../../integration-support.service';
-import { DATA_MAPPER } from '@syndesis/ui/store';
-import { DataShape } from '@syndesis/ui/model';
 import { createBasicFilterModel, findById } from './basic-filter.model';
-import { log, getCategory } from '@syndesis/ui/logging';
 import { BasicFilter } from './filter.interface';
 
 @Component({
