+++ app/ui/src/app/integration/index.ts
@@ -1,4 +1,2 @@
 export * from './integration.module';
-export * from './integration-support.module';
-export * from './integration-support.service';
 export * from './integration.model';
