+++ app/ui/src/app/integration/list/list.component.ts
@@ -13,11 +13,12 @@ import {
   NotificationType
 } from 'patternfly-ng';
 
-import { IntegrationSupportService, Integrations, Integration } from '@syndesis/ui/integration';
-import { IntegrationStore } from '@syndesis/ui/store';
-import { IntegrationViewBase } from '../components';
-import { ModalService, NotificationService } from '@syndesis/ui/common';
 import { log, getCategory } from '@syndesis/ui/logging';
+import { ModalService, NotificationService } from '@syndesis/ui/common';
+import { IntegrationStore } from '@syndesis/ui/store';
+import { IntegrationSupportService } from './../integration-support.service';
+import { Integrations, Integration } from './../integration.model';
+import { IntegrationViewBase } from '../components';
 
 interface ActionConfigs {
   [id: string]: ActionConfig;
