+++ app/ui/src/app/platform/index.ts
@@ -1,3 +1,4 @@
 export * from './guards';
+export * from './helpers';
 export * from './types';
 export * from './validators';
