+++ app/ui/src/app/platform/types/connection/index.ts
@@ -1,0 +1,1 @@
+export * from './connection.models';
