+++ app/ui/src/app/platform/types/user/index.ts
@@ -1,1 +1,2 @@
+export * from './user.models';
 export * from './user.service';
