+++ app/ui/src/app/platform/types/user/user.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 
-import { User } from '@syndesis/ui/model';
+import { User } from './user.models';
 
 @Injectable()
 export abstract class UserService {
