+++ app/ui/src/app/store/action/action.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Restangular } from 'ngx-restangular';
 
+import { Action, Actions } from '@syndesis/ui/platform';
 import { RESTService } from '../entity/rest.service';
-import { Action, Actions } from '../../model';
 
 @Injectable()
 export class ActionService extends RESTService<Action, Actions> {
