+++ app/ui/src/app/store/step/step.store.ts
@@ -1,11 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Step, Steps } from '@syndesis/ui/integration';
-import {
-  Extension,
-  Extensions,
-  Connection,
-  Action
-} from '../../model';
+import { Connection, Action } from '@syndesis/ui/platform';
+import { Extension, Extensions } from '@syndesis/ui/model';
 
 export interface StepKind extends Step {
   name: string;
